Guidelines for dropping project teams
During OpenStack history we have added a lot of OpenStack project
teams, and while we have dropped a few along the way, we do not
have a clear set of considerations to take into account in that
process.

In particular, relying purely on activity levels to consider
dropping projects has proved unproductive in the past, as well
as encouraging people to step up for less-strategic abandoned
projects.

Here is a set of guidelines we should use in future such discussions.

Change-Id: I2c798a6d52a2666657dce201a7c02561d2f4d4ed