Revert "Split OpenDev out of OpenStack Infra"
I think this has merged way too early.  While it was under
project-update, there should be a lot more discussion around
what all of this means for the community, who runs/operates
all the donated infrastructure and a proper governance
before we take such a big action.

We need to have that 'committee' setup, we need to have the
members put in place and the correct infrastructure to then
transition things properly, IMHO.

This reverts commit ae2a861ea3c94a8bb33e1abdee2907423f2aade6.

Change-Id: I66ac74fdda69e2eb21e4766113c27082753a0daa