Split OpenDev out of OpenStack Infra
One of the goals of the OpenDev effort is that it be applicable to
projects that are not OpenStack. While the OpenStack TC and goverance
has worked for us previously, we believe that officially splitting off
OpenDev gives it the freedom it needs to properly address user needs
beyond OpenStack.

Details on OpenDev's future plans can be found here:
http://lists.openstack.org/pipermail/openstack-infra/2019-December/006537.html

Co-Authored-By: Clark Boylan <clark.boylan@gmail.com>
Needed-By: https://review.opendev.org/703488
Change-Id: I512a000720d6402faa3e51551fe4e5841b1231ac