Merge SDK and OSC teams
The teams already share an IRC channel and already share a purpose
in life. In OSC we have a current goal of swapping out client
implementation for SDK, and SDK is already ensuring that it does
what it needs to do to facilitate that goal. The teams also already
share PTG space, and have requested a shared set of time at the
upcoming Denver PTG. Really the separation is historical not
practical at this point, and these days having the additional
layers of governance is not super useful.

Dean has also indicated in IRC that he is not going to be able
to continue working on OSC:

  http://eavesdrop.openstack.org/irclogs/%23openstack-sdks/%23openstack-sdks.2020-03-02.log.html#t2020-03-02T17:59:06

We obviously hope that the universe conspires to change that,
but we all have to accept the sad truth.

Change-Id: Ib1088af1e56e083ee20e8b1623ec7d0fb4335334