Add openstack-tempest-skiplist under TripleO
openstack-tempest-skiplist is a new project, having the collection
of tempest skiplist which is getting moved out from
openstack/tripleo-quickstart-extras repo and can be shared with
TripleO, Openstack Ansible and other openstack projects in
future.

We are keeping it under tripleo group until we find a new group
for the projects shared between TripleO and OpenStack Ansible.

Depends-On: https://review.opendev.org/#/c/713809/

Change-Id: Iac55f8933b6dfca07b91bb3c4f01b66bc88ace0d
Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>