Appoint Yasufumi Ogawa as Tacker PTL
Update Tacker PTL to show my candidancy.
- http://lists.openstack.org/pipermail/openstack-discuss/2020-April/014044.html

I've worked for telcom industry for 17 years as a researcher in NTT
R&D division. I have an experience of maintainer for driving a
sub-project of DPDK which is introduced for realizing high-performance
connectivity between VMs and containers [1]. I've also contributed to
opendev from DPDK side by providing a plugin for the feature and making
feedbacks [2][3].

For Tacker, I've made efforts for bug fixing and reviewing to accelerate
the development for Ussuri release. We will target to refine the
implementation to comply with ESTI standard aiming to become defact, and
to cover container usecases including Kubernetes.

[1] http://git.dpdk.org/apps/spp/
[2] https://opendev.org/x/networking-spp
[3] https://www.openstack.org/summit/vancouver-2018/summit-schedule/speakers/12562

Change-Id: Ib8be004bbf1dab70f7d67c86e59564096b500295