Add nomination for TC chair
I'd like to run for the chair of the technical commitee for the upcoming
cycle, I've served in the past and helped drive discussion and keep up
with the progress of the technical commitee, thanks to the outreach and
help from Doug Hellman at the time.

I plan on continuing to push us to innovate more and drive more changes
through the TC, while simplifying our process.  I intend to go back to
my weekly updates that I would post to our mailing list to keep our
community updated.

I personally also think that there is benefit in being a board member as
well so I can push feedback directly from the TC into the board
meetings.

Change-Id: I03903891fcf83151d6b0010900a382984fca2e81