Add note that OpenStackSDK kept 3.5 support
OpenStackSDK has an important user - OpenDev's Zuul - that's still
supporting and running on 3.5, so OpenStackSDK kept 3.5. Work is
under way to update the OpenDev deployment, and a Zuul v4 is coming
up where it would be appropriate to discuss dropping 3.5 support,
but until that is done SDK would like to continue testing on 3.5.

Change-Id: I379ee9d36332ac9f97fdfc13f8c9d1db463389a5