Infrastructure project team is now the TaCT SIG
Move remaining repositories under the Infrastructure project team
to the Testing and Collaboration Tools SIG.

Technically those repositories are not 'retired' so I did not list
them in legacy.yaml (following the precedent for repositories
previously-owned by Infrastructure and now under the Opendev project).

This change breaks a link to the "community sysadmins" investment
opportunity of 2018/2019, so I apply a quick workaround there.

Depends-On: https://review.opendev.org/721619
Change-Id: I7da9a2060e6cb86fbea0e64a87afa05d46b0a2eb