Loosen voting for community goal proposals
Community goal proposals are intended to be an iterative process
according to the goal process documentation [1].  Goals should be
committed to the goals/proposed/ directory and then the committed goals
should later, as a separate action, be evaluated and the ones selected
as cycle goals should be moved to goals/selected/.

This change clarifies the wording in the goals process and house rules
to indicate that normal voting criteria should hold for proposed goals,
and full formal-vote criteria should hold for the selection of goals.
Presently formal-vote is used for all.

By loosening the criteria for voting it is hoped that the goals section
will be more like the ideas repository, a less constrained forum for
ideas to be proposed, debated, and refined.  Goals would still need at
least two TC members to merge as proposed, so there would still be
validation of the goals.  I believe this is the model that the goals
process refinement was hoping for.

Without this change, the formal-vote criteria for proposed goals voids
the iterative nature of goal development specified in the goal process
document.  Functionally it means that TC members hold off on voting
until goal selection.  This means that goals are not merged to the
goals/proposed/ directory, and there is no reason for it directory to
exist.  I believe this is a suboptimal result.

for more background, see the discussion on #openstack-tc on 28 April
2020 [2].

[1] https://governance.openstack.org/tc/goals/index.html#process-details
[2] http://eavesdrop.openstack.org/irclogs/%23openstack-tc/%23openstack-tc.2020-04-28.log.html#t2020-04-28T16:44:49

Change-Id: Ifa077520468e2caf218e5265f1ce4432ac3848ee