Allow for faster addition of projects
We have a huge timeout right now for the addition of new projects to a
team which is just slowing things down for usually no reason.  This aims
at speeding up the process by replacing the current one-week waiting
period with a need for only two roll-call +1 votes.

We can always revert additions if they don't make sense.

Change-Id: I4a88916f81699a343812c8e01f7419d33cab1ac3