Appoint Javier Peña as rpm-packaging PTL
Javier has accepted to be PTL while we are determining the way
forward for PTL-less projects.

The rpm-packaging doesn't desire to change its practices for
releasing (as it currently owns renderspec and pymod2pkg),
nor does it desire to move to a new repo by being out of
the OpenStack namespaces.

They don't ask much, just to stay alive. (Insert popular song here).

This is a good solution, until we figure out how this team
can become a PTL-less team.

Change-Id: Ia0e6abe463390f8a13757534cdbca79a2d7c58eb