Merging TC and UC into a single body
Create a single governance body to represent all contributors
to OpenStack (developers, operators and end users of the software).

The User Committee is defined in the OpenStack Foundation
bylaws, and amending those is very costly (in time, money and
effort). Luckily, the bylaws defer most of the UC details to a
User Committee charter, which contents are modifiable by the
User Committee itself.

This change adds a minimal User Committee charter that implements
the User Committee as a subcommittee of the Technical Committee.
This allows to operate as a single group without requiring costly
bylaws changes to be implemented first.

Change-Id: I8614f650bbc4176929caf15304cdefaed070c3f6