Remove grace period on resolutions
The charter currently mandates that after a motion receives sufficient
votes to pass, it must stay open for further comments and voting for a
minimum of 3 calendar days. This is in addition to motions needing to
stay open for a minimum of 7 calendar days.

This grace period is a bit impractical to enforce, and in some cases
just delays decisions and creates frustration. This change removes it
from the charter, and advise to apply common sense instead. If the
chair senses that there are concerns that are still unvoiced, a delay
should be applied. If all else fails, it's easy enough to revert a
change if someone disagrees after the fact, using our 'rollback'
house rule[1].

[1] https://governance.openstack.org/tc/reference/house-rules.html

Change-Id: Ib3e12511a8d551ac4fcc76cec79f14b46a704ce0