Record deprecated cycle for deprecated project
Currently project deprecation is done by keeping the
project in project.yaml but with release-management as
deprecated which is not enough to now is it really
deprecated in governance or just release. Also we do not
record the deprecation time so that we can remove the project
after deprecation is over.

Doing it for one project to get agreementon new field and on
documenting it on project guide
- https://review.opendev.org/#/c/736082/

Follow up will be to add the deprecation cycle for all
deprecated repos.
Change-Id: I7c801f29b3d2db14e4b215ad965c111be3f359d4