Add deprecated cycle for deprecated deliverables
Adding the deprecated cycle for deliverables marked as
release-management:deprecated for the known deprecated repo.

Few repo are not so clear if those are actually deprecated
or just release-management is deprecated for them, for example:
devstack-plugin-pika, openstack-ansible-os_karbor etc. They are
left as it is until we have proof or confirmation from the project team
about their deprecation.

Change-Id: I3f3bf5cf65a6ce0e5636c2c1c1687956e8da82c7