Add "tc:approved-release" tag to manila
OpenStack Manila (Shared File Systems as a service)
has been a project under the OpenStack governance
since the Kilo release. The project follows all the
guidelines set forth by the OpenStack Technical
Committee with regard to project structure, development
and releases. It has been increasingly useful in
production OpenStack deployments for a number of
years now.

The "tc:approved-release" tag was introduced in
OpenStack to denote a subset of the projects in the
OpenStack governance that could be used in as
candidates for trademark use under the OpenStack
Foundation trademark policy [1]. At the time that
the tag was introduced, all graduated projects
were added to this list [2]. Manila was not yet
considered a graduated project before the incubation
system of yore gave way to the "big tent" model.

This commit formally requests the "tc:approved-release"
be added to manila, considering the steady growth of the
project, its growing user base, and its consistent
adherence to the TC guidelines.

[1] https://opendev.org/openstack/governance/blame/commit/fb341fe0d9a10acb7662ea507b8b69483e08d21b/reference/tags/tc_approved-release.rst#L17-L19
[2] https://review.opendev.org/198295/

Change-Id: I9beee84153847c36748b0bb47fe0ba519c72cf0f