Add links to chosen release names
We list the release name polling information details in a table, but no
where do we actually provide either the results or links to the results
for someone looking at the release naming process.

This updates the table to link the release letters to the mailing list
announcements declaring the winner of the poll.

Change-Id: Ib583f238134e90828d3be06c047938d7855889ec
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>