[manila] assert:supports-accessible-upgrade
OpenStack Manila offers Shared File Systems
as-a-service to OpenStack users. It has been
around since the Kilo release and has a solid,
long living API which is micro-versioned to
ensure complete backwards compatibility [1].

As a provisioning and a lifecycle management
service for file shares, manila does not place
it self in the data path, thereby, any service
plane changes occurring as a result of upgrade
do not affect resources that it manages.

Manila already has the "assert:supports-upgrade"
tag. A grenade gate job runs on every change
submitted to the project and validates that
controlled resources are not impacted by the
upgrade process [2]. The upgrade process is
documented in the project's administration
guide [3]

[1] https://docs.openstack.org/api-ref/shared-file-system/
[2] https://zuul.opendev.org/t/openstack/builds?project=openstack%2Fmanila&job_name=grenade-dsvm-manila
[3] https://docs.openstack.org/manila/latest/admin/shared-file-systems-upgrades.html

Change-Id: I1ccfb6af91b531c8e1c5236eaa46e344a584889d
Signed-off-by: Goutham Pacha Ravi <gouthampravi@gmail.com>