Sort SIG names in repo owner list
This sorts the list of SIGs in sig-repos to keep it a little more
organized. Also updates some of the names so things actually match
between the names defined in openstack/governance-sigs and the list
here in case we want to start performing any kind of checks or reporting
across the data stored in each location.

Change-Id: Ic1fbe5d7ec5400675ffe7e7de6f69da77ac6ac86
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>