Remove tc:approved-release tag
As per the definition of this tag
"The “OpenStack TC Approved Release” is used as the superset
of projects used by the OpenStack Foundation when creating
commercial trademark programs." (read full definition in [1])

this tag was used to know if project under OpenStack is mature
so that it can be considered as possible candidate for
trademark programs. This situation was in early stage of OpenStack
hen we used to have incubated vs integrated project status but now
everything in project.yaml file is consider as mature, following
release, active etc.

Bylaw requirements are (section 4.1, 4.13[2])
"The Technical Committee shall designate a subset of the OpenStack
Project an “OpenStack Technical Committee Approved Release” from time
to time. The Board of Directors may determine "Trademark Designated
OpenStack Software" from time to time, which will be a subset of the
"OpenStack Technical Committee Approved Release" as provided in
Section 4.1(b)(ii) and (iii)" (read full text in bylaw link[2])

We do not need any changes in bylaw as all projects listed in projects.yaml
file are applicable as "OpenStack Technical Committee Approved Release" and if
any of the current trademarked software is going to be deleted from the
projects.yaml then it need to be done with Coordination Procedures listed
in bylaw.

This commit proposes to remove this tag and add section in place of this
tag defintion to consider all the projects listed
in projects.yaml as "OpenStack Technical Committee Approved Release".

[1] https://governance.openstack.org/tc/reference/tags/tc_approved-release.html#rationale
[2] https://www.openstack.org/legal/bylaws-of-the-openstack-foundation

Change-Id: I0955c20a74eea8cac5f920bba60be1a334d50754