Update projects with Telemetry PTL election result
Due to a bug in the governance update tooling (fixed by
Ia5c477d10787792c80831e2886840a9997e8d1ff which was used to generate
this patch), the results of the Telemetry PTL runoff poll for the
Wallaby cycle were not included in the earlier
I037de61dddf30d9febfea5e819603040e0ccddd2 election results change.
Correctly update the projects list with the result of that one race,
previously missed.

Change-Id: I16ce3758039170c04e92c3c616cd736b5b6530b6