Remove Qinling project team
Qinling did not have any PTL candidate for the Wallaby
cycle. As per its activities in Victoria cycle and no
maintainer, TC decided to retire Qinling deliverables in
the Wallaby cycle. It is announced on ML too[1].

As per retirement process this commit proposes to remove
the Qinling project from the official OpenStack projects.

If there are any maintainer who would like to continue the
development of this projects then they can continue under
opendev in x namespace or repropose it to OpenStack governance.

Depends-On: https://review.opendev.org/c/openstack/project-config/+/764536

[1] http://lists.openstack.org/pipermail/openstack-discuss/2020-November/018638.html

Change-Id: I202c3a880f65ef356ce375062a3b81d087a65c21