Remove Searchlight project team
Searchlight did not have any PTL candidate for the Wallaby
cycle. As per its activities in Victoria cycle and no
maintainer, TC decided to retire Searchlight deliverables in
the Wallaby cycle. It is announced on ML too[1].

As per retirement process this commit proposes to remove
the Searchlight project from the official OpenStack projects.

If there are any maintainer who would like to continue the
development of this projects then they can continue under
opendev in x namespace or repropose it to OpenStack governance.

Depends-On: https://review.opendev.org/c/openstack/project-config/+/764535

[1] http://lists.openstack.org/pipermail/openstack-discuss/2020-November/018637.html

Change-Id: Ia8f0c05c8d6504f889b9abbc3802e55939c947be