Remove Karbor project team
Karbor did not have any PTL candidates for the Wallaby
cycle. As per its activities in Victoria cycle and no
maintainers, the TC decided to retire Karbor deliverables in
the Wallaby cycle. The retirement was announced on the ML[1].

As per retirement process this commit proposes to remove
the Karbor project from the official OpenStack projects.

If there are any maintainers who would like to continue the
development of this projects then they can continue under
opendev in x namespace or repropose it to OpenStack governance.

Thank you to all the contributors of Karbor for your hard work!

[1] http://lists.openstack.org/pipermail/openstack-discuss/2020-November/018643.html

Depends-on: https://review.opendev.org/c/openstack/project-config/+/767057

Change-Id: If4bdd357fea98df9ae7bb94889a5e75c68c9913f