Drop openSUSE from commonly tested distro list
OpenSUSE is no longer maintained for testing
in OpenStack upstream CI, devstack job is broken
for more than month and unfortunately we do not have
any active maintainer to fix and maintain it.

Devstack team is planning to drop the openSUSE job
- https://review.opendev.org/c/openstack/devstack/+/769884

If we cannot test it with working job then we should remove
it from wallaby and future testing runtime.

Agreement on this in TC meeting (14th Jan 2021)
- http://eavesdrop.openstack.org/meetings/tc/2021/tc.2021-01-14-15.00.log.html#l-226

Change-Id: Ia666cc9e306bbea513e94342e92c45ba4d509d42