[manila] add assert:supports-api-interoperability
OpenStack Manila offers a long living API with tests that
ensure backwards compatibility is maintained as the API is
modified.

When evolving the API, the project conforms to the Interoperability
Guideline [1]. The Shared File Systems service API [2] is
micro-versioned [3] and when incrementing the API version, tests
are added to assert compatibility and stability at the prior
micro-version. These tests are never removed until the base version
of the API moves beyond the version that is being tested. So far,
the project has not made any changes to the base API version.

The service provides a version discovery API, conforming to the
version discovery guideline from the API SIG [4]. This version
discovery path can be accessed by unauthenticated users/automation
clients. If a client wants to operate at a prior supported version,
all new API changes remain hidden from the client.

The tests for the API reside in a branchless tempest plugin [5].
The same tempest plugin is used to test all supported stable
branches of the project, including those in extended
maintenance.

[1] https://specs.openstack.org/openstack/api-wg/guidelines/api_interoperability.html
[2] https://docs.openstack.org/api-ref/shared-file-system/
[3] https://specs.openstack.org/openstack/api-wg/guidelines/microversion_specification.html
[4] https://specs.openstack.org/openstack/api-wg/guidelines/microversion_specification.html#version-discovery
[5] https://opendev.org/openstack/manila-tempest-plugin

Change-Id: I557c033629076deb6c3f370123943d904683f85a
Signed-off-by: Goutham Pacha Ravi <gouthampravi@gmail.com>