js-openstack-lib does not make releases
This project was never released yet, was not ready yet for ussuri and victoria
and the OpenstackSDK team think that it require a serious redesign/rewrite [1]
so these changes propose to ignore it from a release point of view.

Set the release-management key so that it gets properly ignored by the release
management team.

[1] http://lists.openstack.org/pipermail/openstack-discuss/2021-January/019853.html

Change-Id: I1a548605a8593cbe2d311a6c2ed307d65cacc93f