Add election schedule in TC chair responsibilities
As per TC charter, PTL and TC election needs to be held
in between of R-6 to R-4 of each release. To comply
that we need to make sure election schedule is prepared
well before time otherwise we may end up adding the
exception for the elections
- https://governance.openstack.org/tc/reference/election-exceptions.html

Having this in TC chair duties can help to have election
on time. I think before R-5 is best time to check with election
WG on schedule preapration.

Change-Id: Ib47681d99b5647f0a9eca57fd338d874c7e7ded3