Replace ATC terminology with AC
This proposes to replace the ATC terminology with AC (Active
Contributor). This makes it clear that ACs are not just people with
technical project contributions. Those contributions are the easiest
way to build the initial list, and automatically qualify a person, but
harder-to-quantify contributions still qualify and we have a procedure
for doing so.

Change-Id: I8ac41cbab493be6a8d58b21234717524edfa70e7