Charter change to handle the vacant seat situation
We had situation in past where we stuck to change the charter
becasue of vacant seat because total number of TC members were
not present to pass the motion.

These type of situation should not hold the TC work and we should
count current number of TC members as complete TC to continue work
on motion, charter etc.

Change-Id: I25c2894a7beb48968edb222cc7ff859fbd392154