Proposing consistent and secure RBAC as a community-wide goal
We have been doing a lot of work towards the secure RBAC for
many cycles. popup team also served its purpose since four
release cycle
- https://wiki.openstack.org/wiki/Consistent_and_Secure_Default_Policies_Popup_Team#Team_Progress

There is a lot of work completed and going on:
- https://review.opendev.org/q/topic:%2522secure-rbac%2522+(status:open+OR+status:merged)

Its time to define this as one of the community-wide goals so that
all remaining projects can finish this work on priority.

Change-Id: Ide5c6d6a58c8550f505b8286a5f7d98218091572