Venus official project status
Venus is a unified log management module, according to the
keyrequirements of OpenStack in log storage, retrieval, analysis
and so on. This module can provide a one-stop solution to log
collection, cleaning, indexing, analysis, alarm, visualization,
report generation etc..

Venus can help operator or maintainer to improve the level of
platform management and the goal of the project is to make log
retrieval no longer difficult, to make log typical error warnings
become a reality.

Venus project has been contributed to OpenStack Community in November
2020. At that time, functions such as collection, retrieval, and error
statistics of platform log and service log  were developed. In order to
make the venus project easier and more practical, the functions of venus
have been enriched and improved in the past six monthsas follows:
1. Improved and standardized the completed code, added some test codes[1].
2. Venus-dashboard plug-in [2] based on Horizon is developed, which can
   provide operators with a log retrieval page, and the function is
   still being improved.
3. To make the deployment more versatile, besides deploy with devstack,
   kolla-based venus deployment script is submit[3].
4. The typical log errors of the openstack platform are being collected,
   and the alarm function is also under development.

The offline promotion of venus is also continuing. Venus is shared on
Hackathon in April 2021, and attracted the interest of companies such as
China Unicom, who express their requirements of Venus. At the upcoming
OpenInfra Days China, Venus will to be promote so that more users and
developers can known and use.

In the future, in addition to improving existing functions, algorithms
will be added to log fault location to enable Venus to play a greater
role in intelligent operation and maintenance.

[1] https://review.opendev.org/q/project:inspur/venus
[2] https://review.opendev.org/q/project:inspur/venus-dashboard
[3] https://review.opendev.org/q/topic:%22venus%22+(status:open%20OR%20status:merged)
[4] ML discussion: http://lists.openstack.org/pipermail/openstack-discuss/2021-January/019748.html

Change-Id: Id82df49a776537bcd38b49a7f95fc0c10bc2f925