Merge 'Technical Writing' SIG into TC
Nowadays, Technical Writing SIG is lacking maintainer and the
current Chair Stephen Finucane would not be able to continue
leading this SIG.

In TC weekly meetings and PTG[1], we decided to merge this
SIG responsibility in TC and a few repos which are related
to upstream institute training will be moved to First Contact
SIG.

One repo 'training-labs' is left for now, which will be called
for the maintainer, and if there is no one, we can retire it.
We are not even sure who is using it. Upstream training and CoA
do not use this repo.

[1] https://etherpad.opendev.org/p/tc-yoga-ptg
    http://lists.openstack.org/pipermail/openstack-discuss/2021-October/025554.html

Needed-By: https://review.opendev.org/c/openstack/governance-sigs/+/815868

Change-Id: Ie745e91e6a0edb8adf6ce5b36dbef22a338d8ca5