Rework the yoga secure RBAC community goal
The Yoga PTG shed a lot of discussion on the useage of system scope and
default roles.

The yoga community goal is still useful, but the discussions highlighted
some confusion in the current approach. We should update the goal to be
consistent with the discussions from the yoga PTG.

This goal is divided into phases spanning the next few development
cycles.

Co-Authored-By: Ghanshyam Mann <gmann@ghanshyammann.com>

Change-Id: I061994e17bd96ace9e3d2040d342b71d3add99cc