Updating the Yoga testing runtime
After the discussion on ML [1], it is not recommended
or best way to drop the py3.6 in Yoga, and keeping it
will help in migration from old (having py3.6) to new
distro (moving to higher python version) versions. For
example, Ubuntu 18.04, CentOS Stream 8 still have 3.6.
CentOS Stream 8 and CentOS Stream 9 both will support
Yoga so we need py3.6 to be tested for a smooth migration.

Considering all these factor to provide a smooth
migration to new python version, TC discussed it in today's
meeting [2] and agreed to keep Python 3.6 testing in Yoga.

We will add CentOS Stream 8 and 9 both with python 3.6 as lowest
and python 3.9 as highest version to run unit tests.

[1] http://lists.openstack.org/pipermail/openstack-discuss/2021-November/025988.html
[2] https://meetings.opendev.org/meetings/tc/2021/tc.2021-12-02-15.02.log.html#l-33

Change-Id: I2d242ac161b19b7615c667f0f7b56826dd1029d8
Co-Authored-By: Radosław Piliszek <radoslaw.piliszek@gmail.com>