Release identification and release name
Since 2010, we went from 26 different OpenStack releases names
(from "A" - "Austin" to "Z" - "Zed"). It's now time to define an
unambiguous and sustainable release schema to easily identify
different OpenStack releases without taking into account the alphabet
iteration.

With this change the TC proposes to move the release identification
from only a release name (ex: "Axxxx") to "year"."release count within
the year" "release name" (ex: 2023.1 Axxxx). This will help to bring
more clarity for releases support and future feature deprecations.

The release name will continue to be used for marketing purposes and
community engagement. The same release name process and criteria will
continue to be used. When the end of the alphabet is reached we start
again from the first letter of the alphabet (after "Z", the next
name should start with "A" again).

The release number of each individual project is independent of the
release identification/name.

Change-Id: I595504a101138ee4a2afd358dc4af8f6617aa845