Define Zed cycle testing runtime
It drop the centos8 stream and py3.6 testing. I did
not add the ubuntu 22.04 as it is not yet released and
will be releases in April but we need to preare the Zed
testing tempate before that. We will discuss it in PTG
and see if we need to update or not(but again based on
its release timming).

Change-Id: I8eaecfd51d5a34701758342dc8cc4e6c522de3bf