Add slaweq nomination as vice-chair
I would like to nominate myself as vice-chair of the
Technical Committee for Zed cycle.

This is my first time ever in TC and I know I have to
learn a lot but I think that being vice-chair and
helping our chair in day-to-day duties will speed up
my ramping up process and I'm positive that I can be
a good vice-chair.

Thanks for considering my nomination.

Change-Id: I171ca2de884b183adf856cbc60d1e9a5591c0183