Retire openstack-helm-docs
As of our discussions from the last PTG, as well as several
prior, one of the goals for the openstack-helm project is to
retire any of the unused repositories under the openstack-helm
projects umbrella.

This change seeks to retire the openstack-helm-docs repo since
it hasn't been maintained and all the documentation lives within
the openstack-helm repo itself.

Depends-on: https://review.opendev.org/c/openstack/openstack-helm-docs/+/839429

Change-Id: Id6d8333c35d27d42c4ac3661971f6c5396b0aa1f