Define Emerging technology and inactive projects framework
At the October 2021 PTG and April 2022 PTG we discussed the need to for
something like "technical preview" state for the e.g. new OpenStack
projects which don't meet all requirements for being official OpenStack
project (see Skyline for example).
This patch adds framework for such projects. It's called "emerging
technology" state which can be set for new projects.
It also defines "inactive" state for the existing, not well maintained
projects.

Change-Id: I32a07f8bb46723f36c1bf846644b6aa768778a4f