Changing new release cadence name
With the outcome of the legal check, it is not clear to us
that tick tock words are ok to use for our new release
cadence arrangement and in what form.

In TC, we decided to use a different name and 'SLURP'
is the choice of the majority[1]

Let's update our resolution document to reflect the new name
and I am adding a "communication" section to explain that how
we will use "SLURP" in our documentation.

[1] https://meetings.opendev.org/meetings/tc/2022/tc.2022-05-12-15.00.log.html#l-191

Change-Id: If82240c1f62842314b48caa4694280c0be147919