Retire openstack-helm-deployments
As part of our discussion from the latest PTG, as well as several
prior, one of the goals for the project was to retire any of the
unused repositories under the openstack-helm umbrella.

This change seeks to retire the openstack-helm-deployments repo
since it has not been maintained and the use-case that it was
originally created for is no longer being actively pursued.

Depends-on: https://review.opendev.org/c/openstack/openstack-helm-deployments/+/847410
Change-Id: Idb63ff8e74427d2b2adab5f003441fb0f6b218b1