Add Adjutant in inactive project list
Adjutant has been inactive since last one year as last
change was merged on Oct 26, 2021 (more than 1 year back)
and their gate is broken. We had PTL volutneer in Zed cycle
but situation is not changed.

TC discussed it in TC weekly meeting[1] and decided to mark it
as Inactive. We will monitor the situation and proceed for the
next step as mentioned in inactive project documentation[2].

[1] https://meetings.opendev.org/meetings/tc/2022/tc.2022-11-30-16.00.log.html#l-173
[2] https://governance.openstack.org/tc/reference/emerging-technology-and-inactive-projects.html#timeline

Change-Id: I5d52497521d4c3ef3dd92388fafa005ed909b36d