Retire openstack-helm-addons
As part of our discussion from the latest PTG, as well as several
prior, one of the goals for the project was to retire any of the
unused repositories under the openstack-helm umbrella.

This change seeks to retire the openstack-helm-addons repo
since it has not been maintained and the existing charts are not
actively supported.

Depends-on: https://review.opendev.org/c/openstack/openstack-helm-addons/+/850890

Change-Id: Iea729e8c591de4d7e1fd680fef575e2a53b1f67f