[goal] Propose to use Ubuntu 22.04 for CI/CD
Ubuntu 22.04 has been released on April 21 and by now already have first
point release as well as upgrade path from 20.04 (Focal). By the end of
Antelope development cycle Ubuntu 22.04 will be around for a year,
which makes a good timing for us to set a goal of migrating CI/CD jobs
to start using this distro.

Change-Id: Ic89ae25dcaebc4f27e9214a857d4d681d0992667