Add Ghanshyam nomination as chair
I would like to nominate myself as chair of the
Technical Committee for next cycle.

I served as TC Chair since Xena cycle and trying my
best to be more active in TC and perform all duties
on time. Few of the things got initiated in the past couple
of cycles and went well (and we will continue these) are:

* TC + Community Leaders interaction: We as TC always
want to get feedback from the community on TC activity and
what direction do they want us to work on? But we hardly
receive any feedback when we initiated it in the past on ML.
Starting a separate session in PTG with TC and Leaders helped
us to receive feedback and also improved the interaction with
community.

* TC doing more Technical work: As feedback from the community
in TC+Leaders sessions, we started doing more technical work than just
project updates (add, remove projects) things. But still, a lot to improve on
this. We will continue working on this and helping the community in both
ways technical as well as from process/governance perspectives.

* Re-starting the interaction with Board members is one of the
good things. It helps to have more visibility of the OpenStack community,
TC as well as discussing our challenges at a broader level.

I would like to continue the chair role and do the same as I am
currently doing. Thanks for considering my nomination.

Change-Id: I8de7cd9ee045e62cfb72b5d7a046bab322f4176b