Make inactive projects timeline as release milestone-2 of the cycle
In current timeline for the inactive projects is the final release
of the cycle which will be little late for the release team to
finish the release work if inactive project become active. Release
milestone-2 is better time to do the check if any inactive projects
needs to be moved to active or retired.

Change-Id: I6ff58fbef50d3fe10f6622330477ada127fcaf18