Deprecate TripleO
As discussed in openstack-discuss[1], TripleO is going to be
deprecated and will be maintained until stable/wallaby.

There are some repo under TripleO project have stable/zed branch
cut and team is not going to maintained it. We discussed this
case in TC meeting and decided[2] to keep stable/zed as 'supported
but no maintainers' (will update this information in stable/zed
README.rst file).

[1] https://lists.openstack.org/pipermail/openstack-discuss/2023-February/032083.html
[2] https://meetings.opendev.org/meetings/tc/2023/tc.2023-03-08-15.59.log.html#l-256

Change-Id: Ifc7c1bea03ab1d4cc339b9f1cf1e1a6c48a76505