Propose Sunbeam as a new OpenStack project
Sunbeam builds on the experience of deploying OpenStack
using the OpenStack Charms project providing a new approach
to the deployment and operation of an OpenStack cloud
which is more contained and lighter-weight in terms of
resource consumption for the OpenStack components.

Juju and Charms remain a core component of this project;
Kubernetes is targetted as a substrate for the majority
of the control plane with data plane components such as
the hypervisor still residing directly on the underlying
servers - a best of both worlds approach.

OCI containers are used for delivery of the K8S workloads
and Snap packages are used for components that are
installed directly on servers.

MicroK8S and Ubuntu are the initial deployment targets
for Sunbeam but due to the containerised approach for
workload deployment it would be possible to target
other Kubernetes distributions and operating systems
in the future if the community wished to move in that
direction.

Currently the Juju charms that support Sunbeam are part of
the OpenStack Charms project but there is no dependency
between the two sets of charms; by placing the Sunbeam
charms and other future project components including
an expanded set of charms, supporting snaps and
documentation under a separate project we make it clear
that this is a distinct, separate effort from the
objectives of the OpenStack Charms project and avoid
confusion for the community around both projects.

The OpenStack Charms project will continue to deliver
support for new OpenStack releases while Sunbeam
incubates into a production grade solution for the
deployment and operation of OpenStack.  An upgrade
path between OpenStack Charm deployed clouds and
Sunbeam is also on the longer term roadmap.

Change-Id: Ia8ae01a073d5fb0ce5f8e4e73a18d0913a49e556