Retire Winstackers project
Winstackers project has no maintainers and the only
company maintaining it was Cloudbase who has changed
their direction and cannot maintain it anymore

- https://lists.openstack.org/pipermail/openstack-discuss/2022-November/031044.html

We have called out for volunteer on ML many times
- https://lists.openstack.org/pipermail/openstack-discuss/2023-March/033035.html
- https://lists.openstack.org/pipermail/openstack-discuss/2023-April/033342.html
- https://lists.openstack.org/pipermail/openstack-announce/2023-April/002069.html

We decided to wait for any volunteer in Vancouver Summit 2023 but no luck.

As there is no volunteer to maintain it, we have to retire
this project and need to remove all the dependencies from OpenStack.

Change-Id: I2a9be90a49294cab72e3627dd097706862254054