Rename extra-atcs key to extra-acs
We have agreed to use term Active Contrubitor rather then
Active Technical Contributor, which means it's high time we've aligned
naming in our tools and releases documentation to align on
term to use. Moreover, it makes even more sense with Extra ACs, since
these are designed for ones, who do not contribute in gerrit but with
other means, which are usually non-technical.

Needed-By: https://review.opendev.org/c/openstack/releases/+/893833
Needed-By: https://review.opendev.org/c/openstack/election/+/893810
Change-Id: Id9f2d182d93a9fbe1357e9ecd4814d49d6eb962f